1.3.0
# New features:

# Realtime Sound File Inferencing.

You are able to load and play sound files, converted to your model's voice in real-time! The way this works is as following:

The sound file replaces your input mic while active. Whatever sound is coming from your loaded file is your "new microphone" while the sound is playing. That means it will infer and play the sound file as if it was your own voice in realtime. You can play speech, singing, or whatever you want. Just make sure the audio is clean, as the client still needs to inference it, no different than the real mic.

When a sound file is playing, it will zero out the input from your real mic, meaning you don't have to worry about overlapping your voice with playback. Mic will automatically unmute when sound is playing again

# Formant Shift:

This works the same as the latest RVC version, shifts voice formant to sound different

# Auto Check for Updates:

Popup when client starts up linking to the latest release page if there is a new version. There is also a check for updates button in the bottom right corner that will flash green if an update is detected.

# Improvements:

Pitch is tunable in 0.1 increments now

Tuned VAD and RNNoise to better clean noise and catch breaths as well as ambient noises. You should leave both on, and also use the built in noise gate (-35 or -40 should be pretty good for catching light noises that get past)

Optimized VAD for real-time usage (Before: 30ms - Now: 5ms ). I cleaned the module to remove extra speech padding calculations that were only used for wav files

# Bug Fixes:

Mute / Unmute label and colors on the button will now be consistent now when muting (Before, sometimes the label and color didn't update based on mute status)